added edit and delete methods fr user role and users

+++ app/Http/Controllers/app/TeamController.php
@@ -14,28 +14,43 @@ class TeamController extends Controller
 {
     public function getUserRoleModal()
     {
-        return view('partials.add_user_role_modal');
+        return view('partials.add_user_role_modal')->with(['title' => "Add Role"]);
+    }
+    public function editUserRole($id)
+    {
+        $user_role = UserRole::find($id);
+        if ($user_role) {
+            return view('partials.add_user_role_modal')->with(['user_role' => $user_role, 'title' => "Update Role"]);
+        }
     }
     public function storeUserRole(Request $req)
     {
-        // $user_role = new UserRole;
-        // $user_role->name = strtoupper($req['name']);
-        // $user_role->save();
-        // return response()->json("Successfully Added");
+        $user_role = new UserRole;
+        $user_role->name = strtoupper($req['name']);
+        $user_role->save();
         return response()->json([
-            "message" => "Success Message",
+            "message" => "User Role successfully Added",
             "location" => "/team"
         ]);
     }
     public function deleteUserRole($id)
     {
         $user_role = UserRole::find($id);
-        $user_role->delete();
+        if ($user_role) {
+            $user_role->delete();
+        } else {
+            return response()->json([
+                'message' => 'User Role Not Found',
+                "location" => "/team"
+            ]);
+        }
         return response()->json([
-            'message' => 'Success',
+            'message' => 'User Role Successfully deleted',
             "location" => "/team"
         ]);
     }
+
+
 
     public function getUserRole()
     {
@@ -51,13 +66,16 @@ class TeamController extends Controller
             'confirm_password' => 'min:3'
         ]);
 
-        $user = $id ? User::find($id) : new User;
+        $user = new User;
         $user->name = $req->name;
         $user->email = $req->email;
         $user->password = Hash::make($req->password);
-        $user->user_role_id = $req->selected_role;
+        $user->user_role_id = $req->user_role;
         $user->save();
-        return response()->json("User Added/updated Successfully");
+        return response()->json([
+            "message" => "User Added successfully",
+            "location" => "/team"
+        ]);
     }
     public function editUserForm($id)
     {
@@ -83,6 +101,9 @@ class TeamController extends Controller
     {
         $permissions = Auth::user()->user_role->permissions()->pluck("name", "id");
         return view('partials.assign_permissions_modal')->with('permissions', $permissions);
+    }
+    public function assignPermissions()
+    {
     }
 
     public function getAllUserRoles()
